split BookmarkCard to its own contained component
